+++ packages/core/src/utils/endpoints.ts
@@ -4,6 +4,7 @@ export const endpoints = {
 	appSettings: '/wp-json/headless-wp/v1/app',
 	category: '/wp-json/wp/v2/categories',
 	tags: '/wp-json/wp/v2/tags',
+	search: '/wp-json/wp/v2/search',
 	tokenVerify: '/wp-json/headless-wp/v1/token',
 	yoast: '/wp-json/yoast/v1/get_head',
 };