+++ wp/headless-wp/includes/classes/BaseToken.php
@@ -41,7 +41,7 @@ abstract class BaseToken {
 	/**
 	 * Return the private key used to encode and decode tokens.
 	 *
-	 * @throws Exception If the private key is not found.
+	 * @throws \Exception If the private key is not found.
 	 *
 	 * @return string
 	 */
@@ -60,7 +60,7 @@ abstract class BaseToken {
 			__( 'Please define either SECURE_AUTH_KEY or TENUP_HEADLESS_JWT_AUTH_KEY in your wp-config.php file.', 'headless-wp' )
 		);
 
-		throw new Exception(
+		throw new \Exception(
 			wp_kses_post(
 				$error->get_error_message()
 			)
@@ -70,9 +70,11 @@ abstract class BaseToken {
 	/**
 	 * Decode capability tokens if present.
 	 *
-	 * @return object
+	 * @param string $token if set will use this instead of reading from the headers
+	 *
+	 * @return null|object
 	 */
-	protected static function get_payload_from_token() {
+	public static function get_payload_from_token( $token = '' ) {
 		// Get HTTP Authorization Header.
 		$header = isset( $_SERVER['HTTP_AUTHORIZATION'] )
 		? sanitize_text_field( wp_unslash( $_SERVER['HTTP_AUTHORIZATION'] ) )
@@ -83,18 +85,25 @@ abstract class BaseToken {
 			$header = sanitize_text_field( wp_unslash( $_SERVER['REDIRECT_HTTP_AUTHORIZATION'] ) );
 		}
 
+		if ( ! $header && isset( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) ) {
+			$header = sanitize_text_field( wp_unslash( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] ) );
+		}
+
 		// No Authorization Header is present.
-		if ( ! $header ) {
+		if ( ! $header && ! $token ) {
 			return null;
 		}
 
 		// Get and parse the token.
 		try {
-			[$token] = sscanf( $header, 'Bearer %s' );
+			if ( ! $token ) {
+				[$token] = sscanf( $header, 'Bearer %s' );
+			}
 
 			if ( empty( $token ) ) {
 				return null;
 			}
+
 			$payload = JWT::decode(
 				$token,
 				self::get_private_key(),