+++ wp/headless-wp/tests/php/includes/TestPreview.php
@@ -7,7 +7,12 @@
 
 namespace HeadlessWP\Tests;
 
+use Exception;
+use DomainException;
 use HeadlessWP\Preview\PreviewLink;
+use HeadlessWP\Preview\PreviewToken;
+use SebastianBergmann\RecursionContext\InvalidArgumentException;
+use PHPUnit\Framework\ExpectationFailedException;
 use Yoast\PHPUnitPolyfills\TestCases\TestCase;
 
 /**
@@ -50,5 +55,55 @@ class TestPreview extends TestCase {
 		$this->assertEquals( $this->preview->handle_preview( 'index.php' ), 'index.php' );
 
 		// todo simulate a preview action
+	}
+
+
+	/**
+  * Tests the preview token generation
+  */
+ public function test_preview_token_generation(): void {
+		$token = PreviewToken::generate( [ 'type' => 'test_token' ] );
+
+		$payload = (array) PreviewToken::get_payload_from_token( $token );
+
+		$this->assertEquals( $payload['type'], 'test_token' );
+	}
+
+	/**
+  * Tests getting the token payload
+  */
+ public function test_get_payload_from_token(): void {
+		$token = PreviewToken::generate( [ 'type' => 'test_token' ] );
+
+		$_SERVER['HTTP_AUTHORIZATION'] = 'Bearer ' . $token;
+
+		$payload = (array) PreviewToken::get_payload_from_token();
+
+		$this->assertEquals( $payload['type'], 'test_token' );
+
+		unset( $_SERVER['HTTP_AUTHORIZATION'] );
+
+		$_SERVER['REDIRECT_HTTP_AUTHORIZATION'] = 'Bearer ' . $token;
+
+		$payload = (array) PreviewToken::get_payload_from_token();
+
+		$this->assertEquals( $payload['type'], 'test_token' );
+
+		unset( $_SERVER['REDIRECT_HTTP_AUTHORIZATION'] );
+	}
+
+	/**
+  * Tests getting the token payload from alternative header
+  */
+ public function test_get_payload_from_token_alternative_header(): void {
+		$token = PreviewToken::generate( [ 'type' => 'test_token' ] );
+
+		$_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] = 'Bearer ' . $token;
+
+		$payload = (array) PreviewToken::get_payload_from_token();
+
+		$this->assertEquals( $payload['type'], 'test_token' );
+
+		unset( $_SERVER['HTTP_X_HEADSTARTWP_AUTHORIZATION'] );
 	}
 }